Add files via upload
This project is an interactive electronic game designed to test a player's focus and motor skills. The objective is to move a metal wand through a zig-zag wire maze without making contact. If the wand touches the maze, the player loses a life. To make it more exciting, if all lives are lost, the game assigns a random dare to the player before restarting.