Add files via upload
This project is an automated lighting system based on the Arduino platform. It utilizes a Light Dependent Resistor (LDR) to detect environmental light intensity. The system automatically turns on an LED when the surroundings become dark and turns it off when there is sufficient light, simulating the behavior of smart streetlights or garden lamps